+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add `auto_propagate` flag to Peer [#57]
+- Add optional `height` parameter to `broadcast` method [#57]
+
 ### Changed
 
 - Change `on_message` callback into a trait [#63]
@@ -27,5 +30,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - RaptorQ as Forward Error Correction.
 - Examples in `example` dir
 
+[#57]: https://github.com/dusk-network/kadcast/issues/57
 [#60]: https://github.com/dusk-network/kadcast/issues/60
-[#60]: https://github.com/dusk-network/kadcast/issues/63
+[#63]: https://github.com/dusk-network/kadcast/issues/63